+++ lib/default_http_header_plugin.dart
@@ -11,13 +11,18 @@ class DefaultHttpHeaderPlugin {
     return version;
   }
 
+  static Future<Map<dynamic, dynamic>> get allDefaultHeaders async {
+    final headers = await _channel.invokeMethod('getAllDefaultHeaders');
+    return headers;
+  }
+
   static Future<String> get acceptLanguage async {
     final String acceptLanguage = await _channel.invokeMethod('getAcceptLanguage');
     return acceptLanguage;
   }
 
-  static Future<Map<dynamic, dynamic>> get allDefaultHeaders async {
-    final headers = await _channel.invokeMethod('getAllDefaultHeaders');
-    return headers;
+  static Future<String> get userAgent async {
+    final String userAgent = await _channel.invokeMethod('getUserAgent');
+    return userAgent;
   }
 }