+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.1+1
+
+* Add user-agent header.
+
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Add accept-language header.